Don't upgrade RubyGems in Ruby3.2
Other Ruby versions are not updating RubyGems from what comes
from the Lambda base image.

The upgrade is causing an issue with the just released 4.0.0 where
the `--without` flag has been removed. That flag is used by Lambda
Builders to build Ruby projects, so updating breaks that flow.

This needs to be fixed in Lambda Builders before updating the
version of RubyGems anyway. For now we're good because the base
image hasn't updated to the latest version.